+++ src/main/java/spharos/client/worker/infrastructure/WorkerRepository.java
@@ -1,8 +1,7 @@
-package spharos.client.service.infrastructure;
+package spharos.client.worker.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import spharos.client.service.domain.worker.Worker;
-
+import spharos.client.worker.domain.Worker;
 import java.util.List;
 
 public interface WorkerRepository extends JpaRepository<Worker, Long> {
