+++ src/main/java/spharos/client/worker/presentation/WorkerController.java
@@ -1,0 +1,14 @@
+package spharos.client.worker.presentation;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/client/worker")
+@Slf4j
+public class WorkerController {
+
+}
